Add `enable` to `livenessProbe` and `readinessProbe`
The default values where fine for most of the LLMs, but the bigger ones
as e.g. Llama 3.1 405B were failing because the `initialDelaySeconds`
was not enough as the download for those takes > 20 minutes, and there
was no way of modifying those before. Now we're keeping the same
sensible defaults, but still adding a note and allowing updates on those
values depending on which LLM is being deployed with HUGS.