Release 0.0.3 version of HUGS Helm Chart (#1)
* Update `charts/hugs/templates/tests` to use `--retry`

* Add `enable` to `livenessProbe` and `readinessProbe`

The default values where fine for most of the LLMs, but the bigger ones
as e.g. Llama 3.1 405B were failing because the `initialDelaySeconds`
was not enough as the download for those takes > 20 minutes, and there
was no way of modifying those before. Now we're keeping the same
sensible defaults, but still adding a note and allowing updates on those
values depending on which LLM is being deployed with HUGS.

* Rename `image.model` to `image.name`

* Align `values.yaml` and `aws/eks-values.yaml`

* Update `version` in `charts/hugs/{Chart.yaml,index.yaml}`

* Update `README.md`, `aws/README.md` and `charts/hugs/templates/_helpers.tpl`

Missing udpates from version bump and "model" rename to "name"

* Remove `AWS_MARKETPLACE_PRODUCT_CODE`

* Update to use `tag: 0.1.0` as sensible default

* Update `charts/hugs/index.yaml`